Unify Python e2e tests and add CI coverage
