+++ .github/workflows/ci.yml
@@ -263,9 +263,10 @@ jobs:
           pnpm --dir sdk/js typecheck
           pnpm --dir sdk/js test
 
-  a2a_compat:
-    name: A2A Compatibility
+  python_e2e:
+    name: Python E2E
     runs-on: ubuntu-latest
+    timeout-minutes: 20
     env:
       TESTCONTAINERS_RYUK_DISABLED: "true"
     steps:
@@ -287,14 +288,26 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y --no-install-recommends pkg-config libssl-dev protobuf-compiler
 
+      - name: Check Docker daemon
+        run: docker info
+
       - name: Build Talon binaries
         run: cargo build --locked --bins
 
       - name: Install Python test dependencies
         run: python -m pip install -r tests/requirements.txt
 
-      - name: Run upstream A2A SDK compatibility test
-        run: pytest -q tests/test_a2a_compat.py
+      - name: Run Python E2E suite
+        run: |
+          pytest -q \
+            tests/test_a2a_compat.py \
+            tests/test_workers.py \
+            tests/test_chat.py \
+            tests/test_search.py \
+            tests/test_resources.py \
+            tests/test_workflows.py \
+            tests/test_cli_sqlite.py \
+            tests/test_acp_sqlite.py
 
   aws_local_e2e:
     name: AWS Local E2E
