+++ tests/test_acp_sqlite.py
@@ -1,286 +1,32 @@
-import grpc
 import json
 import os
-import pytest
 import shutil
 import subprocess
-import threading
 import time
 import uuid
 
-from talon_client import (
-    TalonClient,
-    CreateNamespaceRequest,
-    CreateResourceRequest,
-    CreateSessionRequest,
-    GetSessionRequest,
-    ListResourcesRequest,
-    SendMessageRequest,
-    StreamSessionPartsRequest,
-)
-from talon_client.resources import AgentSpec, Model, ResourceManifest, ResourceMeta, ResourceSpec
+import pytest
+
 import conftest
-from e2e.cli_harness import TalonCli
+from e2e.blackbox import last_assistant_message, message_text
 from e2e import scenarios as e2e
-
-PART_TYPE_TEXT = 1
-PART_TYPE_REASONING = 2
-PART_TYPE_USAGE = 5
+from e2e.stack import E2EStack
+from talon_client import GetSessionRequest, ListResourcesRequest
 
 
-def message_text(message):
-    return "".join(part.content for part in message.parts if part.part_type == PART_TYPE_TEXT)
+@pytest.fixture
+def stack(sqlite_local_stack: E2EStack) -> E2EStack:
+    return sqlite_local_stack
 
 
-def assistant_messages(messages):
-    return [message for message in messages if message.role == 2]
-
-
-def last_assistant_message(messages):
-    assistants = assistant_messages(messages)
-    return assistants[-1] if assistants else None
-
-
-def ensure_namespace(stub, name):
-    try:
-        stub.namespaces.Create(CreateNamespaceRequest(name=name, recursive=True))
-    except grpc.RpcError as err:
-        if err.code() != grpc.StatusCode.ALREADY_EXISTS:
-            raise
-
-
-def create_agent_resource(stub, ns, name, spec):
-    return stub.resources.Create(
-        CreateResourceRequest(
-            ns=ns,
-            manifest=ResourceManifest(
-                api_version="talon.impalasys.com/v1",
-                kind="Agent",
-                metadata=ResourceMeta(name=name, namespace=ns),
-                spec=ResourceSpec(agent=spec),
-            ),
-        )
-    ).resource
-
-
-def wait_for_worker_endpoint(stub, expected_url, attempts=30, delay=1):
-    last_urls = []
-    for _ in range(attempts):
-        resources = list(stub.resources.List(ListResourcesRequest(ns="Sys", kind="Worker")).resources)
-        last_urls = [
-            endpoint.url
-            for resource in resources
-            if resource.status and resource.status.worker.phase == "ready"
-            for endpoint in resource.status.worker.endpoints
-        ]
-        if expected_url in last_urls:
-            return
-        time.sleep(delay)
-    raise AssertionError(
-        f"Timed out waiting for worker endpoint {expected_url!r}; saw {last_urls!r}"
-    )
-
-
-def test_worker_registers_sqlite_local_endpoint(
-    gateway_channel_sqlite,
-    talon_infrastructure_sqlite,
-):
-    stub = TalonClient(gateway_channel_sqlite)
-    wait_for_worker_endpoint(stub, "http://127.0.0.1:18082")
-
-
-def test_single_turn_chat_sqlite_local_socket(gateway_channel_sqlite, mock_llm_server):
-    stub = TalonClient(gateway_channel_sqlite)
-
-    ensure_namespace(stub, "talon-sqlite-test")
-
-    agent = create_agent_resource(
-        stub,
-        "talon-sqlite-test",
-        "test-llm-agent",
-        AgentSpec(
-                model_policy={
-                    "profiles": [
-                        {
-                            "name": "default",
-                            "model": Model(
-                                provider="mock",
-                                name="minimax-m2.7",
-                                temperature=0.7,
-                            ),
-                        }
-                    ]
-                },
-                system_prompt="You are a helpful test assistant.",
-            ),
-    )
-    assert agent.metadata.name == "test-llm-agent"
-
-    session = stub.sessions.Create(
-        CreateSessionRequest(agent="test-llm-agent", ns="talon-sqlite-test")
-    )
-    session_id = session.session_id
-    assert session_id != ""
-
-    stub.sessions.SendMessage(
-        SendMessageRequest(
-            agent="test-llm-agent",
-            session_id=session_id,
-            ns="talon-sqlite-test",
-            message="What is the square root of 144?",
-        )
-    )
-
-    success = False
-    messages = []
-    for _ in range(30):
-        time.sleep(1)
-        res = stub.sessions.Get(
-            GetSessionRequest(
-                agent="test-llm-agent",
-                session_id=session_id,
-                ns="talon-sqlite-test",
-            )
-        )
-        messages = res.messages
-        assistant = last_assistant_message(messages)
-        if res.state == "IDLE" and assistant is not None:
-            success = True
-            break
-
-    assert success, "Agent did not reply in time or failed to revert to IDLE"
-    agent_message = last_assistant_message(messages)
-    assert agent_message is not None
-    assert agent_message.role == 2
-    assert "12" in message_text(agent_message)
-
-
-def test_streaming_chat_sqlite_local_socket(gateway_channel_sqlite, mock_llm_server):
-    stub = TalonClient(gateway_channel_sqlite)
-
-    ensure_namespace(stub, "talon-sqlite-stream-test")
-
-    create_agent_resource(
-        stub,
-        "talon-sqlite-stream-test",
-        "stream-agent",
-        AgentSpec(
-                model_policy={
-                    "profiles": [
-                        {
-                            "name": "default",
-                            "model": Model(
-                                provider="mock",
-                                name="minimax",
-                                temperature=0.7,
-                            ),
-                        }
-                    ]
-                },
-                system_prompt="Stream me.",
-            ),
-    )
-
-    session = stub.sessions.Create(
-        CreateSessionRequest(agent="stream-agent", ns="talon-sqlite-stream-test")
-    )
-    session_id = session.session_id
-
-    def send_msg():
-        time.sleep(2.0)
-        stub.sessions.SendMessage(
-            SendMessageRequest(
-                agent="stream-agent",
-                session_id=session_id,
-                ns="talon-sqlite-stream-test",
-                message="Stream test message",
-            )
-        )
-
-    sender = threading.Thread(target=send_msg)
-    sender.start()
-
-    stream_req = StreamSessionPartsRequest(
-        agent="stream-agent",
-        session_id=session_id,
-        ns="talon-sqlite-stream-test",
-    )
-    events = []
-    try:
-        saw_reasoning = False
-        saw_token = False
-        saw_usage = False
-        for idx, event in enumerate(stub.sessions.StreamParts(stream_req)):
-            events.append(event)
-            if event.part.part_type == PART_TYPE_REASONING:
-                saw_reasoning = True
-            if event.part.part_type == PART_TYPE_TEXT:
-                saw_token = True
-            if event.part.part_type == PART_TYPE_USAGE:
-                saw_usage = True
-            if saw_reasoning and saw_token and saw_usage:
-                break
-            if idx > 20:
-                break
-    except grpc.RpcError:
-        pass
-    sender.join()
-
-    assert len(events) >= 1
-    reasoning_events = [event for event in events if event.part.part_type == PART_TYPE_REASONING]
-    token_events = [event for event in events if event.part.part_type == PART_TYPE_TEXT]
-    usage_events = [event for event in events if event.part.part_type == PART_TYPE_USAGE]
-    assert len(reasoning_events) >= 1
-    assert len(token_events) >= 1
-    assert len(usage_events) >= 1
-    assert "Inspecting the request" in reasoning_events[0].part.content
-    streamed_text = "".join(event.part.content for event in token_events)
-    assert "received" in streamed_text
-
-
-def apply_manifest_with_cli(path, grpc_port):
-    cli = conftest.get_binary_path("talon_cli")
-    auth_args = []
-    api_key = os.environ.get(conftest.api_key_env_name(grpc_port))
-    if api_key:
-        auth_args.extend(["--api-key", api_key])
-    result = subprocess.run(
-        [
-            cli,
-            "--gateway",
-            f"http://127.0.0.1:{grpc_port}",
-            *auth_args,
-            "apply",
-            "-f",
-            str(path),
-        ],
-        text=True,
-        capture_output=True,
-        check=False,
-    )
-    assert result.returncode == 0, (
-        f"talon-cli apply failed for {path}\n"
-        f"stdout:\n{result.stdout}\n"
-        f"stderr:\n{result.stderr}"
-    )
-    return result.stdout
-
-
-def cli_for_grpc_port(grpc_port):
-    env = {}
-    auth_file = os.environ.get(conftest.api_key_auth_file_env_name(grpc_port))
-    if auth_file:
-        env["TALON_AUTH_FILE"] = auth_file
-    return TalonCli(
-        conftest.get_binary_path("talon_cli"),
-        f"http://127.0.0.1:{grpc_port}",
-        api_key=os.environ.get(conftest.api_key_env_name(grpc_port)),
-        env=env,
-    )
-
-
-def acp_vars(suffix, source_ns, target_ns, deployment_name, **extra):
-    values = {
+def acp_vars(
+    suffix: str,
+    source_ns: str,
+    target_ns: str,
+    deployment_name: str,
+    **extra: str,
+) -> dict[str, str]:
+    values: dict[str, str] = {
         "run_id": suffix,
         "source_ns": source_ns,
         "target_ns": target_ns,
@@ -290,155 +36,14 @@ def acp_vars(suffix, source_ns, target_ns, deployment_name, **extra):
     return values
 
 
-def apply_acp_stack(cli, manifest_names, variables):
+def apply_acp_stack(
+    cli,
+    manifest_names: list[str],
+    variables: dict[str, str],
+) -> None:
     manifest_dir = e2e.MANIFEST_ROOT / "acp"
     for name in manifest_names:
         e2e.apply(cli, manifest_dir / name, variables)
-
-
-def test_cli_chat_sqlite_local_socket(sqlite_test_grpc_port, mock_llm_server):
-    cli = cli_for_grpc_port(sqlite_test_grpc_port)
-    suffix = uuid.uuid4().hex[:8]
-    namespace = f"talon-cli-chat-{suffix}"
-    agent = "cli-chat-agent"
-    e2e.apply(
-        cli,
-        e2e.MANIFEST_ROOT / "chat" / "agent.yaml",
-        {
-            "namespace": namespace,
-            "agent": agent,
-            "system_prompt": "You are a helpful CLI E2E test assistant.",
-        },
-    )
-    session_id = e2e.session_create(cli, namespace, agent)
-    e2e.session_send(
-        cli,
-        namespace,
-        agent,
-        session_id,
-        "What is the square root of 144?",
-    )
-    completed = e2e.wait_for_session_text(cli, namespace, agent, session_id, "12")
-    assert completed["state"] == "IDLE"
-
-
-def test_cli_filtered_search_sqlite_local_socket(sqlite_test_grpc_port, mock_llm_server):
-    cli = cli_for_grpc_port(sqlite_test_grpc_port)
-    suffix = uuid.uuid4().hex[:8]
-    namespace = f"talon-cli-search-{suffix}"
-    agent = "cli-search-agent"
-    token = f"clisearchtoken{suffix}"
-    e2e.apply(
-        cli,
-        e2e.MANIFEST_ROOT / "chat" / "agent.yaml",
-        {
-            "namespace": namespace,
-            "agent": agent,
-            "system_prompt": "You are a helpful search CLI E2E test assistant.",
-        },
-    )
-    session_id = e2e.session_create(cli, namespace, agent)
-    e2e.session_send(
-        cli,
-        namespace,
-        agent,
-        session_id,
-        f"Please index {token} for CLI search.",
-    )
-
-    last_output = ""
-    for _ in range(30):
-        result = cli.run(
-            "search",
-            "sessions",
-            "--namespace",
-            namespace,
-            "--agent",
-            agent,
-            token,
-        )
-        last_output = result.stdout
-        if token in last_output:
-            break
-        time.sleep(1)
-
-    assert token in last_output
-    assert namespace in last_output
-    assert "SessionMessage" in last_output
-    assert "part" in last_output
-
-
-def test_cli_streaming_chat_sqlite_local_socket(sqlite_test_grpc_port, mock_llm_server):
-    cli = cli_for_grpc_port(sqlite_test_grpc_port)
-    suffix = uuid.uuid4().hex[:8]
-    namespace = f"talon-cli-stream-{suffix}"
-    agent = "cli-stream-agent"
-    e2e.apply(
-        cli,
-        e2e.MANIFEST_ROOT / "chat" / "agent.yaml",
-        {
-            "namespace": namespace,
-            "agent": agent,
-            "system_prompt": "You are a helpful streaming CLI E2E test assistant.",
-        },
-    )
-
-    session_id = e2e.session_create(cli, namespace, agent)
-    events = e2e.session_send_stream_json(
-        cli,
-        namespace,
-        agent,
-        session_id,
-        "Stream test message",
-    )
-    parts = [
-        event.get("part") or {}
-        for event in events
-        if event.get("part") is not None
-    ]
-    reasoning = [part for part in parts if part.get("type") == "reasoning"]
-    text = [part for part in parts if part.get("type") == "text"]
-    usage = [part for part in parts if part.get("type") == "usage"]
-
-    assert reasoning
-    assert text
-    assert usage
-    assert "Inspecting the request" in reasoning[0].get("content", "")
-    assert "received" in "".join(part.get("content", "") for part in text)
-
-
-def test_cli_apply_rejects_status_in_resource_manifest(sqlite_test_grpc_port, tmp_path):
-    manifest = tmp_path / "agent-with-status.yaml"
-    manifest.write_text(
-        """
-apiVersion: talon.impalasys.com/v1
-kind: Agent
-metadata:
-  name: status-owned
-  namespace: status-owned-test
-spec:
-  systemPrompt: This manifest should fail before apply.
-status:
-  phase: Ready
-"""
-    )
-    cli = conftest.get_binary_path("talon_cli")
-    result = subprocess.run(
-        [
-            cli,
-            "--gateway",
-            f"http://127.0.0.1:{sqlite_test_grpc_port}",
-            "apply",
-            "-f",
-            str(manifest),
-        ],
-        text=True,
-        capture_output=True,
-        check=False,
-    )
-
-    assert result.returncode != 0
-    assert "Resource manifests cannot set status" in (result.stderr + result.stdout)
 
 
 def list_resources(stub, namespace, kind):
@@ -650,20 +255,6 @@ def wait_for_sandbox_process(stub, namespace):
     )
 
 
-def wait_for_sandbox(stub, namespace):
-    last_count = 0
-    for _ in range(60):
-        sandboxes = list_resources(stub, namespace, "Sandbox")
-        last_count = len(sandboxes)
-        if sandboxes:
-            return sandboxes[0]
-        time.sleep(1)
-    raise AssertionError(
-        f"Timed out waiting for sandbox in namespace {namespace}; "
-        f"last sandbox count={last_count}"
-    )
-
-
 def wait_for_cli_sandbox(cli, namespace, attempts=60):
     last_count = 0
     for _ in range(attempts):
@@ -695,9 +286,11 @@ def wait_for_cli_sandbox_process(cli, namespace, attempts=30):
 
 
 def test_cli_apply_acp_deployment_starts_session_sqlite_local_socket(
-    sqlite_test_grpc_port,
-):
-    cli = cli_for_grpc_port(sqlite_test_grpc_port)
+    stack: E2EStack,
+) -> None:
+    # Verify the ACP deployment templates render correctly, create the expected
+    # runtime resources, and can complete a permission-gated coding session.
+    cli = stack.cli()
     suffix = uuid.uuid4().hex[:8]
     source_ns = f"customers-{suffix}"
     target_ns = f"{source_ns}:acme"
@@ -855,8 +448,10 @@ def test_cli_apply_acp_deployment_starts_session_sqlite_local_socket(
 
 
 def test_cli_apply_live_acp_deployment_runs_code_in_sandbox_sqlite_local_socket(
-    sqlite_test_grpc_port,
-):
+    stack: E2EStack,
+) -> None:
+    # Verify the selected live ACP Docker harness can create its resources, run
+    # code inside the sandbox, and return the expected completion signal.
     if not live_acp_e2e_enabled():
         pytest.skip(
             "set TALON_ACP_DOCKER_E2E=1 to run the live ACP harness Docker e2e"
@@ -864,7 +459,7 @@ def test_cli_apply_live_acp_deployment_runs_code_in_sandbox_sqlite_local_socket(
     harness = live_acp_harness_config()
     ensure_docker_acp_image(harness["image"])
 
-    cli = cli_for_grpc_port(sqlite_test_grpc_port)
+    cli = stack.cli()
     suffix = uuid.uuid4().hex[:8]
     harness_slug = harness["harness"].replace("-", "_")
     source_ns = f"{harness['harness']}-customers-{suffix}"
