+++ tests/test_search.py
@@ -1,145 +1,68 @@
 import os
 import sys
-import time
 import uuid
-
-import grpc
 
 # Important: Add generated protos to path so "proto.xxx" resolves locally and not to proto_plus
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "generated")))
 
 from e2e import scenarios as e2e
+from e2e.blackbox import (
+    append_user_message,
+    assert_session_message_document,
+    create_agent_resource,
+    create_resource,
+    ensure_namespace,
+)
+from e2e.stack import E2EStack
 from talon_client import (
-    AppendSessionMessageRequest,
-    CreateNamespaceRequest,
-    CreateResourceRequest,
     CreateSessionRequest,
     DeleteSessionRequest,
     GetSearchResultRequest,
     TalonClient,
 )
-from talon_client.data import (
-    ROLE_USER,
-    SESSION_MESSAGE_PART_TYPE_TEXT,
-    SessionMessage,
-    SessionMessagePart,
-)
-from talon_client.resources import (
-    AgentSpec,
-    KnowledgeSpec,
-    Model,
-    ResourceManifest,
-    ResourceMeta,
-    ResourceSpec,
-    WorkflowSpec,
-)
+from talon_client.resources import AgentSpec, KnowledgeSpec, Model, ResourceSpec, WorkflowSpec
 
 
-def ensure_namespace(stub, name):
-    try:
-        stub.namespaces.Create(CreateNamespaceRequest(name=name, recursive=True))
-    except grpc.RpcError as err:
-        if err.code() != grpc.StatusCode.ALREADY_EXISTS:
-            raise
-
-
-def create_resource(stub, ns, kind, name, spec):
-    return stub.resources.Create(
-        CreateResourceRequest(
-            ns=ns,
-            manifest=ResourceManifest(
-                api_version="talon.impalasys.com/v1",
-                kind=kind,
-                metadata=ResourceMeta(name=name, namespace=ns),
-                spec=spec,
-            ),
-        )
-    ).resource
-
-
-def create_agent(stub, namespace, name):
-    return create_resource(
-        stub,
-        namespace,
-        "Agent",
-        name,
-        ResourceSpec(
-            agent=AgentSpec(
-                model_policy={
-                    "profiles": [
-                        {
-                            "name": "default",
-                            "model": Model(
-                                provider="mock",
-                                name="minimax/minimax-m2.7",
-                                temperature=0.0,
-                            ),
-                        }
-                    ]
-                },
-                system_prompt="Search E2E test agent.",
-            )
-        ),
-    )
-
-
-def append_user_message(stub, namespace, agent, session_id, text, token):
-    now = int(time.time() * 1_000_000)
-    message_id = f"msg-{token}"
-    return stub.sessions.AppendMessage(
-        AppendSessionMessageRequest(
-            ns=namespace,
-            agent=agent,
-            session_id=session_id,
-            message=SessionMessage(
-                id=message_id,
-                role=ROLE_USER,
-                created_at=now,
-                labels={"source": "search-e2e"},
-                parts=[
-                    SessionMessagePart(
-                        id="000000",
-                        part_type=SESSION_MESSAGE_PART_TYPE_TEXT,
-                        content=text,
-                        created_at=now,
-                    )
-                ],
-            ),
-        )
-    ).message
-
-
-def assert_session_message_document(result, namespace, agent, session_id, token):
-    document = result.document
-    assert document.source.namespace == namespace
-    assert document.source.kind == "SessionMessage"
-    assert document.document_kind == "part"
-    assert document.attributes.get("part_type", "") == "TEXT"
-    assert document.attributes.get("agent", "") == agent
-    assert document.attributes.get("session_id", "") == session_id
-    assert token in result.snippet
-
-
-def test_postgres_pubsub_session_search_indexes_opens_and_deletes(
-    gateway_channel,
-    mock_llm_server,
-):
-    stub = TalonClient(gateway_channel)
+def test_session_search_indexes_opens_and_deletes(
+    stack: E2EStack,
+    client: TalonClient,
+) -> None:
+    # Verify that session message search indexes a user message, can open the
+    # indexed document content, and removes the document after session deletion.
     suffix = uuid.uuid4().hex[:8]
-    namespace = f"talon-search-{suffix}"
+    namespace = f"talon-search-{stack.name}-{suffix}"
     agent = f"search-agent-{suffix}"
     token = f"sessiontoken{suffix}"
-    message_text = f"Please remember {token} for session search."
+    expected_message_text = f"Please remember {token} for session search."
 
-    ensure_namespace(stub, namespace)
-    create_agent(stub, namespace, agent)
-    session_id = stub.sessions.Create(
+    ensure_namespace(client, namespace)
+    create_agent_resource(
+        client,
+        namespace,
+        agent,
+        AgentSpec(
+            model_policy={
+                "profiles": [
+                    {
+                        "name": "default",
+                        "model": Model(
+                            provider="mock",
+                            name="minimax/minimax-m2.7",
+                            temperature=0.0,
+                        ),
+                    }
+                ]
+            },
+            system_prompt="Search E2E test agent.",
+        ),
+    )
+    session_id = client.sessions.Create(
         CreateSessionRequest(ns=namespace, agent=agent)
     ).session_id
-    append_user_message(stub, namespace, agent, session_id, message_text, token)
+    append_user_message(client, namespace, agent, session_id, expected_message_text, token)
 
     result, _ = e2e.wait_for_session_search(
-        stub,
+        client,
         namespace,
         agent,
         token,
@@ -149,18 +72,18 @@ def test_postgres_pubsub_session_search_indexes_opens_and_deletes(
     )
     assert_session_message_document(result, namespace, agent, session_id, token)
 
-    opened = stub.searches.GetResult(
+    opened = client.searches.GetResult(
         GetSearchResultRequest(ns=namespace, document_id=result.document.id)
     )
     assert opened.document.ref.id == result.document.id
-    assert message_text in opened.content
+    assert expected_message_text in opened.content
 
-    deleted = stub.sessions.Delete(
+    deleted = client.sessions.Delete(
         DeleteSessionRequest(ns=namespace, agent=agent, session_id=session_id)
     )
     assert deleted.success is True
     e2e.wait_for_no_session_search_results(
-        stub,
+        client,
         namespace,
         agent,
         token,
@@ -168,23 +91,43 @@ def test_postgres_pubsub_session_search_indexes_opens_and_deletes(
     )
 
 
-def test_postgres_pubsub_workspace_search_indexes_resources_and_knowledge(
-    gateway_channel,
-    mock_llm_server,
-):
-    stub = TalonClient(gateway_channel)
+def test_workspace_search_indexes_resources_and_knowledge(
+    stack: E2EStack,
+    client: TalonClient,
+) -> None:
+    # Verify that workspace search indexes both workflow metadata and knowledge
+    # content and returns the stored document bodies when results are opened.
     suffix = uuid.uuid4().hex[:8]
-    namespace = f"talon-workspace-search-{suffix}"
+    namespace = f"talon-workspace-search-{stack.name}-{suffix}"
     agent = f"workspace-agent-{suffix}"
     workflow_name = f"workflow-{suffix}"
     knowledge_name = f"knowledge-{suffix}"
     workflow_token = f"workflowtoken{suffix}"
     knowledge_token = f"knowledgetoken{suffix}"
 
-    ensure_namespace(stub, namespace)
-    create_agent(stub, namespace, agent)
+    ensure_namespace(client, namespace)
+    create_agent_resource(
+        client,
+        namespace,
+        agent,
+        AgentSpec(
+            model_policy={
+                "profiles": [
+                    {
+                        "name": "default",
+                        "model": Model(
+                            provider="mock",
+                            name="minimax/minimax-m2.7",
+                            temperature=0.0,
+                        ),
+                    }
+                ]
+            },
+            system_prompt="Search E2E test agent.",
+        ),
+    )
     create_resource(
-        stub,
+        client,
         namespace,
         "Workflow",
         workflow_name,
@@ -196,7 +139,7 @@ def test_postgres_pubsub_workspace_search_indexes_resources_and_knowledge(
     )
 
     workflow_result, _ = e2e.wait_for_search(
-        stub,
+        client,
         namespace,
         workflow_token,
         lambda item: item.document.source.kind == "Workflow"
@@ -209,14 +152,14 @@ def test_postgres_pubsub_workspace_search_indexes_resources_and_knowledge(
     assert workflow_doc.document_kind == "metadata"
     assert workflow_token in workflow_result.snippet
 
-    opened_workflow = stub.searches.GetResult(
+    opened_workflow = client.searches.GetResult(
         GetSearchResultRequest(ns=namespace, document_id=workflow_doc.id)
     )
     assert workflow_name in opened_workflow.content
     assert workflow_token in opened_workflow.content
 
     create_resource(
-        stub,
+        client,
         namespace,
         "Knowledge",
         knowledge_name,
@@ -229,7 +172,7 @@ def test_postgres_pubsub_workspace_search_indexes_resources_and_knowledge(
     )
 
     knowledge_result, _ = e2e.wait_for_search(
-        stub,
+        client,
         namespace,
         knowledge_token,
         lambda item: item.document.source.kind == "Knowledge"
@@ -242,43 +185,7 @@ def test_postgres_pubsub_workspace_search_indexes_resources_and_knowledge(
     assert knowledge_doc.document_kind == "content"
     assert knowledge_token in knowledge_result.snippet
 
-    opened_knowledge = stub.searches.GetResult(
+    opened_knowledge = client.searches.GetResult(
         GetSearchResultRequest(ns=namespace, document_id=knowledge_doc.id)
     )
     assert knowledge_token in opened_knowledge.content
-
-
-def test_sqlite_local_socket_session_search_indexes_and_opens(
-    gateway_channel_sqlite,
-    mock_llm_server,
-):
-    stub = TalonClient(gateway_channel_sqlite)
-    suffix = uuid.uuid4().hex[:8]
-    namespace = f"talon-sqlite-search-{suffix}"
-    agent = f"sqlite-search-agent-{suffix}"
-    token = f"sqlitesessiontoken{suffix}"
-    message_text = f"SQLite local search should find {token}."
-
-    ensure_namespace(stub, namespace)
-    create_agent(stub, namespace, agent)
-    session_id = stub.sessions.Create(
-        CreateSessionRequest(ns=namespace, agent=agent)
-    ).session_id
-    append_user_message(stub, namespace, agent, session_id, message_text, token)
-
-    result, _ = e2e.wait_for_session_search(
-        stub,
-        namespace,
-        agent,
-        token,
-        lambda item: item.document.attributes.get("session_id", "") == session_id
-        and item.document.document_kind == "part",
-        session_id=session_id,
-    )
-    assert_session_message_document(result, namespace, agent, session_id, token)
-
-    opened = stub.searches.GetResult(
-        GetSearchResultRequest(ns=namespace, document_id=result.document.id)
-    )
-    assert opened.document.ref.id == result.document.id
-    assert message_text in opened.content
